Handle GG, TG, SV subset cards
SWSH series introduced subsets where numbers overlap. In deck lists,
there is no way to separate these because CRZ 10 could refer to
Mew (swsh12pt5gg-gg10) or Kriketot (swsh12pt5-10).

The mappings are precomputed so that whenever one of these special sets
are encountered, we can use the card name to look up whatever the proper
card ID is. If it's not in there, then it's in the original set and
falls back to the regular algorithm.